New : linear mode

+++ Farfadet.h
@@ -14,7 +14,7 @@ public:
   void init(uint8_t txPin, int stepsPerTurn, Stream& serial);
 
   void setControlMode(uint8_t mode);
-  void setTargetPosition(long target);//in degres or in meters depending on the mode
+  void setTargetPosition(float target);//in degres or in meters depending on the mode
   long getCurrentPosition();
   void setTargetSpeed(float speed);
   float getCurrentSpeed();
